+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a new completely system to run the simulation and collisions of everything.
 - Added textured imposters for planets.
 - Fixed issues when rendering imposters.
+- Fixed an error that rendered starfields using small scale matrices.
 
 ## 0.0